Add support for exporting flow information in NetFlow v5 format.
This is implemented by having the datapath send a new meta-Flow End message
that contains all the information needed by NetFlow v5 and the OpenFlow
Flow Expiration messages.  secchan grabs these Flow End messages and
generates any requested Flow End and NetFlow messages.  The Flow End
message is implemented as a Nicira vendor extension, but it is only used
internally between the datapath and secchan, so the switch is still fully
compatible with OpenFlow v0.8.9.

NOTE: This change has not been ported to "switch", which means that it is
not able to generate NetFlow messages.  "switch" is no longer maintained
and will be removed from the repository on January 1, 2009.