Fix learning-switch STP breakage from "out_port" in flow stats request.
ofp_flow_stats_request recently added a new member, "out_port", to select
only flows that output to a particular port.  Unfortunately this code
in learning-switch.c was not updated to set that member to OFPP_NONE,
with the result that it would only get flows that output to port 0.

This bug was found when looking at this code for another reason and thus
is no guarantee that the STP code in learning-switch actually works.