xenserver: Don't include veth_mod.ko in RPMs.
veth_mod.ko is built only for Linux 2.6.18 (since later versions already
have it).  Our XenServer build doesn't use it at all, so don't package it.

(This is in response to a build failure against a XenServer 5.7.0
prerelease, which uses a 2.6.27 kernel and thus for which veth_mod.ko is
not built.)